+++ src/main/kotlin/ru/hh/plugins/geminio/sdk/recipe/parsers/widgets/WidgetsSectionParser.kt
@@ -11,23 +11,31 @@ private const val KEY_WIDGETS_SECTION = "widgets"
 
 private const val KEY_STRING_PARAMETER_TYPE = "stringParameter"
 private const val KEY_BOOLEAN_PARAMETER_TYPE = "booleanParameter"
+private const val KEY_SUGGEST_PARAMETER_TYPE = "suggestParameter"
 
 /**
  * Parser from YAML to [ru.hh.plugins.geminio.sdk.recipe.models.widgets.WidgetsSection].
  */
-internal fun Map<String, Any>.toWidgetsSection(): WidgetsSection {
+internal fun Map<String, Any>.toWidgetsSection(
+    recipeRootDirPath: String? = null,
+): WidgetsSection {
     val widgetsList = requireNotNull(this[KEY_WIDGETS_SECTION] as? List<Map<String, Any>>) {
         rootSectionErrorMessage(KEY_WIDGETS_SECTION)
     }
 
     return WidgetsSection(
-        parameters = widgetsList.map { it.toRecipeParameter() }
+        parameters = widgetsList.map { widget ->
+            widget.toRecipeParameter(recipeRootDirPath)
+        }
     )
 }
 
-private fun Map<String, Any>.toRecipeParameter(): RecipeParameter {
+private fun Map<String, Any>.toRecipeParameter(
+    recipeRootDirPath: String?,
+): RecipeParameter {
     val stringParameterMap = this[KEY_STRING_PARAMETER_TYPE] as? Map<String, Any>
     val booleanParameterMap = this[KEY_BOOLEAN_PARAMETER_TYPE] as? Map<String, Any>
+    val suggestParameterMap = this[KEY_SUGGEST_PARAMETER_TYPE] as? Map<String, Any>
 
     return when {
         stringParameterMap != null -> {
@@ -38,6 +46,13 @@ private fun Map<String, Any>.toRecipeParameter(): RecipeParameter {
             booleanParameterMap.toWidgetsBooleanParameter("$KEY_WIDGETS_SECTION:$KEY_BOOLEAN_PARAMETER_TYPE")
         }
 
+        suggestParameterMap != null -> {
+            suggestParameterMap.toWidgetsSuggestParameter(
+                sectionName = "$KEY_WIDGETS_SECTION:$KEY_SUGGEST_PARAMETER_TYPE",
+                recipeRootDirPath = recipeRootDirPath,
+            )
+        }
+
         else -> {
             throw IllegalArgumentException(
                 sectionUnknownEnumKeyErrorMessage(
@@ -46,6 +61,7 @@ private fun Map<String, Any>.toRecipeParameter(): RecipeParameter {
                     listOf(
                         KEY_STRING_PARAMETER_TYPE,
                         KEY_BOOLEAN_PARAMETER_TYPE,
+                        KEY_SUGGEST_PARAMETER_TYPE,
                     ).joinToString { "'$it'" }
                 )
             )